Refactored the Numbers and Colors Activities with fragments
