+++ app/src/main/java/com/example/android/miwok/ColorsActivity.java
@@ -1,162 +1,16 @@
 package com.example.android.miwok;
 
-import android.app.Activity;
-import android.content.Context;
-import android.media.AudioManager;
-import android.media.MediaPlayer;
-import android.support.v4.app.NavUtils;
-import android.support.v4.content.ContextCompat;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.util.Log;
-import android.view.MenuItem;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.AdapterView;
-import android.widget.LinearLayout;
-import android.widget.ListView;
-
-import java.util.ArrayList;
 
 public class ColorsActivity extends AppCompatActivity {
-
-    private static final String TAG = "ColorsActivity";
-    //Create audio object
-    private MediaPlayer mMediaPlayer;
-
-    //Handles audio focus when playing a sound file
-    private AudioManager mAudioManager;
-
-    AudioManager.OnAudioFocusChangeListener mOnAudioFocusChangeListener = new AudioManager.OnAudioFocusChangeListener() {
-
-        public void onAudioFocusChange(int focusChange) {
-            if (focusChange == AudioManager.AUDIOFOCUS_LOSS_TRANSIENT || focusChange == AudioManager.AUDIOFOCUS_LOSS_TRANSIENT_CAN_DUCK ) {
-                // Pause playback
-                mMediaPlayer.pause();
-                mMediaPlayer.seekTo(0);
-            } else if (focusChange == AudioManager.AUDIOFOCUS_GAIN) {
-                // The AUDIOFOCUS_GAIN state means we have regained focus and can resume playback
-                mMediaPlayer.start();
-            } else if (focusChange == AudioManager.AUDIOFOCUS_LOSS) {
-                //The AUDIOFOCUS_LOSS case means we've lost audio focus and can stop playback and cleanup resources
-                releaseMediaPlayer();
-            }
-        }
-    };
-
-    /*
-    * This listener gets triggered when the {@link Media Player} has completed playing the audio file
-    * */
-    private MediaPlayer.OnCompletionListener mCompletionListener = new MediaPlayer.OnCompletionListener() {
-        @Override
-        public void onCompletion(MediaPlayer mediaPlayer) {
-            releaseMediaPlayer();
-        }
-    };
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.word_list);
-        //Adds a left-facing caret alongside the app icon and enables it as an action button
-        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-
-        // Create a list of words
-        ArrayList<Word> colors = new ArrayList<Word>();
-        colors.add(new Word("red", "weṭeṭṭi", R.drawable.color_red, R.raw.color_red));
-        colors.add(new Word("green", "chokokki", R.drawable.color_green, R.raw.color_green));
-        colors.add(new Word("brown", "ṭakaakki", R.drawable.color_brown, R.raw.color_brown));
-        colors.add(new Word("gray", "ṭopoppi", R.drawable.color_gray, R.raw.color_gray));
-        colors.add(new Word("black", "kululli", R.drawable.color_black, R.raw.color_black));
-        colors.add(new Word("white", "kelelli", R.drawable.color_white, R.raw.color_white));
-        colors.add(new Word("dusty yellow", "ṭopiisә", R.drawable.color_dusty_yellow, R.raw.color_dusty_yellow));
-        colors.add(new Word("mustard yellow", "chiwiiṭә", R.drawable.color_mustard_yellow, R.raw.color_mustard_yellow));
-
-        // Create an {@link WordAdapter}, whose data source is a list of {@link Word}s. The
-        // adapter knows how to create list items for each item in the list.
-        WordAdapter adapter = new WordAdapter(this, colors, R.color.category_colors);
-
-        // Find the {@link ListView} object in the view hierarchy of the {@link Activity}.
-        // There should be a {@link ListView} with the view ID called list, which is declared in the
-        // activity_numbers.xml layout file.
-        final ListView listView = (ListView) findViewById(R.id.list);
-
-        // Make the {@link ListView} use the {@link WordAdapter} we created above, so that the
-        // {@link ListView} will display list items for each {@link Word} in the list.
-        listView.setAdapter(adapter);
-
-        // Create and setup the {@link AudioManager} to request audio focus
-        mAudioManager = (AudioManager) ColorsActivity.this.getSystemService(Context.AUDIO_SERVICE);
-
-        // Create a clickListener to play the audio file
-        listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
-            @Override
-            public void onItemClick (AdapterView < ? > adapter, View view, int position, long arg){
-                /*
-                * Store the audio resource ID to be passed to the mMediaPlayer variable*/
-                Word audioResource = (Word)listView.getItemAtPosition(position);
-
-                //Clear media player
-                releaseMediaPlayer();
-
-                // Create and setup the {@link AudioManager} to request audio focus
-                int result = mAudioManager.requestAudioFocus(mOnAudioFocusChangeListener,
-                        // Use the music stream.
-                        AudioManager.STREAM_MUSIC,
-                        // Request permanent focus.
-                        AudioManager.AUDIOFOCUS_GAIN_TRANSIENT);
-                if (result == AudioManager.AUDIOFOCUS_REQUEST_GRANTED) {
-                    // We have audio focus now.
-                    mMediaPlayer = MediaPlayer.create(ColorsActivity.this, audioResource.getAudioResourceId());
-
-                    // Start audio playback.
-                    mMediaPlayer.start(); // no need to call prepare(); create() does that for you
-
-                    //Release the Media Player once the audio playback is complete
-                    mMediaPlayer.setOnCompletionListener(mCompletionListener);
-                }
-            }
-        });
-    }
-
-    @Override
-    protected void onStop() {
-        super.onStop();
-        /*
-        * When the activity is stopped, release the media player resources
-        * because we won't be playing anymore sounds*/
-        releaseMediaPlayer();
-    }
-
-    //    Backbutton functionality
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        switch (item.getItemId()) {
-            // Respond to the action bar's Up/Home button
-            case android.R.id.home:
-                NavUtils.navigateUpFromSameTask(this);
-                return true;
-        }
-        return super.onOptionsItemSelected(item);
-    }
-    /**
-     * Clean up the media player by releasing its resources.
-     */
-    private void releaseMediaPlayer() {
-        // If the media player is not null, then it may be currently playing a sound.
-        if (mMediaPlayer != null) {
-            // Regardless of the current state of the media player, release its resources
-            // because we no longer need it.
-            mMediaPlayer.release();
-
-            // Set the media player back to null. For our code, we've decided that
-            // setting the media player to null is an easy way to tell that the media player
-            // is not configured to play an audio file at the moment.
-            mMediaPlayer = null;
-
-            //Abandon audio focus when playback is complete
-            mAudioManager.abandonAudioFocus(mOnAudioFocusChangeListener);
-            Log.d(TAG, "MediaPlayer released");
-        }
+        setContentView(R.layout.activity_category);
+        getSupportFragmentManager().beginTransaction()
+                .replace(R.id.container, new ColorsFragment())
+                .commit();
     }
 }
