+++ app/src/main/java/com/example/android/miwok/ColorsFragment.java
@@ -1,0 +1,158 @@
+package com.example.android.miwok;
+
+
+import android.content.Context;
+import android.media.AudioManager;
+import android.media.MediaPlayer;
+import android.os.Bundle;
+import android.support.v4.app.Fragment;
+import android.util.Log;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.AdapterView;
+import android.widget.ListView;
+import android.widget.TextView;
+
+import java.util.ArrayList;
+
+/**
+ * A simple {@link Fragment} subclass.
+ */
+public class ColorsFragment extends Fragment {
+
+    private static final String TAG = "ColorsActivity";
+    //Create audio object
+    private MediaPlayer mMediaPlayer;
+
+    //Handles audio focus when playing a sound file
+    private AudioManager mAudioManager;
+
+    AudioManager.OnAudioFocusChangeListener mOnAudioFocusChangeListener = new AudioManager.OnAudioFocusChangeListener() {
+
+        public void onAudioFocusChange(int focusChange) {
+            if (focusChange == AudioManager.AUDIOFOCUS_LOSS_TRANSIENT || focusChange == AudioManager.AUDIOFOCUS_LOSS_TRANSIENT_CAN_DUCK ) {
+                // Pause playback
+                mMediaPlayer.pause();
+                mMediaPlayer.seekTo(0);
+            } else if (focusChange == AudioManager.AUDIOFOCUS_GAIN) {
+                // The AUDIOFOCUS_GAIN state means we have regained focus and can resume playback
+                mMediaPlayer.start();
+            } else if (focusChange == AudioManager.AUDIOFOCUS_LOSS) {
+                //The AUDIOFOCUS_LOSS case means we've lost audio focus and can stop playback and cleanup resources
+                releaseMediaPlayer();
+            }
+        }
+    };
+
+    /*
+    * This listener gets triggered when the {@link Media Player} has completed playing the audio file
+    * */
+    private MediaPlayer.OnCompletionListener mCompletionListener = new MediaPlayer.OnCompletionListener() {
+        @Override
+        public void onCompletion(MediaPlayer mediaPlayer) {
+            releaseMediaPlayer();
+        }
+    };
+
+
+
+    /**
+     * Clean up the media player by releasing its resources.
+     */
+    private void releaseMediaPlayer() {
+        // If the media player is not null, then it may be currently playing a sound.
+        if (mMediaPlayer != null) {
+            // Regardless of the current state of the media player, release its resources
+            // because we no longer need it.
+            mMediaPlayer.release();
+
+            // Set the media player back to null. For our code, we've decided that
+            // setting the media player to null is an easy way to tell that the media player
+            // is not configured to play an audio file at the moment.
+            mMediaPlayer = null;
+
+            //Abandon audio focus when playback is complete
+            mAudioManager.abandonAudioFocus(mOnAudioFocusChangeListener);
+            Log.d(TAG, "MediaPlayer released");
+        }
+    }
+
+    @Override
+    public void onStop() {
+        super.onStop();
+        // When the activity is stopped, release the media player resources because we won't
+        // be playing any more sounds.
+        releaseMediaPlayer();
+    }
+    public ColorsFragment() {
+        // Required empty public constructor
+    }
+
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+
+        View rootView = inflater.inflate(R.layout.word_list, container, false);
+
+        // Create a list of words
+        ArrayList<Word> colors = new ArrayList<Word>();
+        colors.add(new Word("red", "weṭeṭṭi", R.drawable.color_red, R.raw.color_red));
+        colors.add(new Word("green", "chokokki", R.drawable.color_green, R.raw.color_green));
+        colors.add(new Word("brown", "ṭakaakki", R.drawable.color_brown, R.raw.color_brown));
+        colors.add(new Word("gray", "ṭopoppi", R.drawable.color_gray, R.raw.color_gray));
+        colors.add(new Word("black", "kululli", R.drawable.color_black, R.raw.color_black));
+        colors.add(new Word("white", "kelelli", R.drawable.color_white, R.raw.color_white));
+        colors.add(new Word("dusty yellow", "ṭopiisә", R.drawable.color_dusty_yellow, R.raw.color_dusty_yellow));
+        colors.add(new Word("mustard yellow", "chiwiiṭә", R.drawable.color_mustard_yellow, R.raw.color_mustard_yellow));
+
+        // Create an {@link WordAdapter}, whose data source is a list of {@link Word}s. The
+        // adapter knows how to create list items for each item in the list.
+        WordAdapter adapter = new WordAdapter(getActivity(), colors, R.color.category_numbers);
+
+        // Find the {@link ListView} object in the view hierarchy of the {@link Activity}.
+        // There should be a {@link ListView} with the view ID called list, which is declared in the
+        // activity_numbers.xml layout file.
+        final ListView listView = (ListView) rootView.findViewById(R.id.list);
+
+        // Make the {@link ListView} use the {@link WordAdapter} we created above, so that the
+        // {@link ListView} will display list items for each {@link Word} in the list.
+        listView.setAdapter(adapter);
+
+        // Create and setup the {@link AudioManager} to request audio focus
+        mAudioManager = (AudioManager) getActivity().getSystemService(Context.AUDIO_SERVICE);
+
+        // Create a clickListener to play the audio file
+        listView.setOnItemClickListener(new AdapterView.OnItemClickListener() {
+            @Override
+            public void onItemClick (AdapterView < ? > adapter, View view, int position, long arg){
+                /*
+                * Store the audio resource ID to be passed to the mMediaPlayer variable*/
+                Word audioResource = (Word)listView.getItemAtPosition(position);
+
+                //Clear media player
+                releaseMediaPlayer();
+
+                // Create and setup the {@link AudioManager} to request audio focus
+                int result = mAudioManager.requestAudioFocus(mOnAudioFocusChangeListener,
+                        // Use the music stream.
+                        AudioManager.STREAM_MUSIC,
+                        // Request permanent focus.
+                        AudioManager.AUDIOFOCUS_GAIN_TRANSIENT);
+                if (result == AudioManager.AUDIOFOCUS_REQUEST_GRANTED) {
+                    // We have audio focus now.
+                    mMediaPlayer = MediaPlayer.create(getActivity(), audioResource.getAudioResourceId());
+
+                    // Start audio playback.
+                    mMediaPlayer.start(); // no need to call prepare(); create() does that for you
+
+                    //Release the Media Player once the audio playback is complete
+                    mMediaPlayer.setOnCompletionListener(mCompletionListener);
+                }
+            }
+        });
+
+        return rootView;
+    }
+}
